--- a/mvcc/kvstore.go
+++ b/mvcc/kvstore.go
@@ -463,6 +463,7 @@ func (s *store) restore() error {
 			scheduledCompact = 0
 		}
 	}
+
 	tx.Unlock()
 
 	if scheduledCompact != 0 {