#5317 

@xiang90 Currently i have commented the test cases. Will fix all the test cases and also add the benchmark test case for this in this PR. There are too many test cases dependent on this. Will start fixing them one by one.
Thank you!