@heyitsanthony 2016-11-14
Please don't comment out so many tests, it makes the PR difficult to read. Most (e.g., the e2e ones) should work as-is if the fix is good...