@sinsharat 2016-11-14
@heyitsanthony its still initial phase will be un-commenting all these cases in this PR itself after fixing them all. Or shall i remove all these commented code and keep it as original itself?