@sinsharat 2016-11-14
@heyitsanthony as you suggested will keep only the changes here. Once the all test cases have been fixed, this PR should automatically pass.