@sinsharat 2016-11-16
Benchmark results:

![benchmark_sore_restore](https://cloud.githubusercontent.com/assets/20169156/20351915/29fef33c-ac3b-11e6-881a-6311b904a186.png)