@xiang90 2016-11-16
@sinsharat Do you have the result before this PR?