@sinsharat 2016-11-16
@xiang90 No i don't have the result. I had done the changes based on you suggestion. If you want will take the benchmark result for the older one.