@sinsharat 2016-11-16
@xiang90 The bench test case i had written was not working with old code. So i had to modify the code as below:
```
st := time.Now()
s = NewStore(be, &lease.FakeLessor{}, &i)
b.Logf("Took %v for restoring %v keys\n", time.Since(st), b.N)
```
After modifying the bench test, the following were the readings with old and new code:
**Old Code:**
```
BenchmarkStoreRestoreRevs10-4             	   50000	     25233 ns/op
--- BENCH: BenchmarkStoreRestoreRevs10-4
	kvstore_bench_test.go:117: Took 43.911ms for restoring 10 keys
	kvstore_bench_test.go:117: Took 67.0137ms for restoring 200 keys
	kvstore_bench_test.go:117: Took 150.127ms for restoring 2000 keys
	kvstore_bench_test.go:117: Took 81.0395ms for restoring 20000 keys
	kvstore_bench_test.go:117: Took 563.2987ms for restoring 200000 keys
	kvstore_bench_test.go:117: Took 1.1537436s for restoring 500000 keys
```

**New Code:**
```
BenchmarkStoreRestoreRevs10-4             	  100000	     17500 ns/op
--- BENCH: BenchmarkStoreRestoreRevs10-4
	kvstore_bench_test.go:117: Took 58.8606ms for restoring 10 keys
	kvstore_bench_test.go:117: Took 93.2865ms for restoring 100 keys
	kvstore_bench_test.go:117: Took 96.471ms for restoring 1000 keys
	kvstore_bench_test.go:117: Took 81.7109ms for restoring 10000 keys
	kvstore_bench_test.go:117: Took 412.4868ms for restoring 100000 keys
	kvstore_bench_test.go:117: Took 566.0508ms for restoring 300000 keys
	kvstore_bench_test.go:117: Took 1.5504294s for restoring 1000000 keys
```