@xiang90 2016-11-16
@sinsharat Can you also try 1 rev per key case?