@sinsharat 2016-11-16
@xiang90 Please find below the result for 1 rev per key case:
**Old Code:**
```
BenchmarkStoreRestoreRevs1-4              	  200000	      5922 ns/op
--- BENCH: BenchmarkStoreRestoreRevs1-4
	kvstore_bench_test.go:116: Took 43.6651ms for restoring 1 keys
	kvstore_bench_test.go:116: Took 76.5179ms for restoring 10 keys
	kvstore_bench_test.go:116: Took 64.1562ms for restoring 100 keys
	kvstore_bench_test.go:116: Took 138.4179ms for restoring 1000 keys
	kvstore_bench_test.go:116: Took 134.2355ms for restoring 10000 keys
	kvstore_bench_test.go:116: Took 584.6538ms for restoring 100000 keys
	kvstore_bench_test.go:116: Took 1.0278762s for restoring 200000 keys
```

**New Code:**
```
BenchmarkStoreRestoreRevs1-4              	  200000	      5472 ns/op
--- BENCH: BenchmarkStoreRestoreRevs1-4
	kvstore_bench_test.go:116: Took 61.997ms for restoring 1 keys
	kvstore_bench_test.go:116: Took 57.2608ms for restoring 10 keys
	kvstore_bench_test.go:116: Took 66.9507ms for restoring 100 keys
	kvstore_bench_test.go:116: Took 105.1183ms for restoring 1000 keys
	kvstore_bench_test.go:116: Took 119.1674ms for restoring 10000 keys
	kvstore_bench_test.go:116: Took 541.6184ms for restoring 100000 keys
	kvstore_bench_test.go:116: Took 998.8248ms for restoring 200000 keys
```