@xiang90 2016-11-16
@sinsharat OK. I thought 1 rev case should be slower. But it seems that the overhead is minimum.

Does these result make sense to you? I will give this a closer look today or early tomorrow.