@sinsharat 2016-11-16
@xiang90 yes what i can figure out is that the new code, for smaller number of keys takes slight;y more time, but as the keys to restore grows, the restoration becomes faster than the old code. Also if you see the operation time, overall its quite less for newer code for both 10rev/key(17500 ns/op) and 1rev/key(5472 ns/op) as opposed to older code having10rev/key(25233 ns/op) and 1rev/key(5922 ns/op).