@sinsharat 2016-11-16
@xiang90 I have done the changes as per your suggestion. PTAL.
Thank you!